+++ client/src/App.tsx
@@ -1,15 +1,19 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
+import { ThemeProvider } from '@/shared/context/ThemeContext'
 import Layout from './components/Layout'
 import DesignPlaygroundPage from './pages/DesignPlaygroundPage'
 import LandingPage from './pages/LandingPage'
 import MainPage from './pages/MainPage'
-import AboutPage from './pages/AboutPage'
 import PortfolioPage from './pages/PortfolioPage'
-import BlogPage from './pages/BlogPage'
+import LearningPage from './pages/LearningPage'
+import ColumnPage from './pages/ColumnPage'
+import ProjectPage from './pages/ProjectPage'
+import AboutPage from './pages/AboutPage'
 
 function App() {
   return (
-    <BrowserRouter>
+    <ThemeProvider>
+      <BrowserRouter>
       <Routes>
         <Route path="/design-playground" element={<DesignPlaygroundPage />} />
         <Route path="/" element={<LandingPage />} />
@@ -17,10 +21,13 @@ function App() {
           <Route path="main" element={<MainPage />} />
           <Route path="about" element={<AboutPage />} />
           <Route path="portfolio" element={<PortfolioPage />} />
-          <Route path="blog" element={<BlogPage />} />
+          <Route path="learning" element={<LearningPage />} />
+          <Route path="column" element={<ColumnPage />} />
+          <Route path="project" element={<ProjectPage />} />
         </Route>
       </Routes>
     </BrowserRouter>
+    </ThemeProvider>
   )
 }